% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCGM-package.R
\docType{package}
\name{cleanCGM-package}
\alias{cleanCGM-package}
\title{cleanCGM: Clean CGM Signal}
\description{
Set of functions to clean CGM signal as collected in the EXTREME and the PREFIT-UP projects. Expected input files are csv or xlsx format with various file structures and missing values expected. The output is RData files including a data frame with two columns, i.e., timestamp and glucose in mg/dL.
}
\keyword{internal}
