% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sample_data}
\alias{sample_data}
\title{Sample CGM data files}
\description{
Raw CGM data in various formats for testing and demonstration.
}
\section{Files}{

\itemize{
\item \code{freestylelibre_csv_spanish.csv}: FreeStyle Libre device standard CSV export in Spanish.
\item \code{freestylelibre_txt_spanish.txt}: FreeStyle Libre device tab-separated file with Spanish column names.
\item \code{freestylelibre_xlsx_spanish.xlsx}: FreeStyle Libre device excel export with metadata header and Spanish column names.
}
}

