% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGlucose.R
\name{formatGlucose}
\alias{formatGlucose}
\title{Formats glucose in a consistent manner}
\usage{
formatGlucose(data, glucCol = NULL)
}
\arguments{
\item{data}{data frame read by readCGM}

\item{glucCol}{Character. Name of the column with the glucose information}
}
\value{
Glucose formatted
}
\description{
Formats glucose in a consistent manner
}
