% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTime.R
\name{formatTime}
\alias{formatTime}
\title{Formats timestamp in a consistent manner}
\usage{
formatTime(data, timeCol = NULL)
}
\arguments{
\item{data}{data frame read by readCGM}

\item{timeCol}{Character. Name of the column with the timestamp.}
}
\value{
timestamp formatted
}
\description{
Formats timestamp in a consistent manner
}
