% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCGM.R
\name{readCGM}
\alias{readCGM}
\title{Reads file with CGM data either in .txt or .xlsx format}
\usage{
readCGM(files)
}
\arguments{
\item{file}{Character. Path to file with CGM data (accepts txt and xlsx formats).}
}
\value{
Data frame with CGM data.
}
\description{
Reads file with CGM data either in .txt or .xlsx format
}
